keep mayan fees in contract

+++ src/interfaces/IMctpDriver.sol
@@ -5,9 +5,9 @@ pragma solidity ^0.8.0;
 interface IMctpDriver {
 	function mctpSwap(
 		address tokenIn,
-		uint64 amountIn,
+		uint256 amountIn,
 		address tokenOut,
-		uint64 promisedAmountOut,
-		uint64 gasDrop
+		uint256 promisedAmountOut,
+		uint256 gasDrop
 	) external;
 }
